+++ libshvbroker/include/shv/broker/azureconfig.h
