+++ libshvbroker/include/shv/broker/clientconnectionnode.h
