+++ libshvbroker/include/shv/broker/currentclientshvnode.h
