+++ libshvbroker/include/shv/broker/groupmapping.h
