+++ libshvbroker/include/shv/broker/shvbrokerglobal.h
