+++ libshvbroker/src/brokerapp.cpp
@@ -1,15 +1,15 @@
-#include "brokerapp.h"
 #include "aclmanagersqlite.h"
-#include "currentclientshvnode.h"
-#include "clientshvnode.h"
-#include "brokerappnode.h"
-#include "subscriptionsnode.h"
 #include "brokeraclnode.h"
-#include "clientconnectionnode.h"
+#include "brokerappnode.h"
 #include "brokerrootnode.h"
+#include "clientshvnode.h"
 #include "rpc/brokertcpserver.h"
 #include "rpc/clientconnectiononbroker.h"
 #include "rpc/masterbrokerconnection.h"
+#include "subscriptionsnode.h"
+#include <shv/broker/brokerapp.h>
+#include <shv/broker/currentclientshvnode.h>
+#include <shv/broker/clientconnectionnode.h>
 
 #ifdef WITH_SHV_WEBSOCKETS
 #include "rpc/websocketserver.h"
